1.19 Fix issues regarding the size of the poster based on device size
This commit fixes an issues regarding the size of the poster on some
devices. The poster size that will be downloaded from TheMovieDb is
determined by the pixel density of the device & not enough options
were added originally. This resulted in some devices having posters
that were too big or too small.