1.20 Add complete README.md
This commit adds a full README for this project. The readme includes
a brief description of the project. It contains instructions on how
to install the project on your system. It has a brief description of
my code design decisions for the project with future updates. Finally,
it has a standard Apache 2.0 license.